Changing from the R native `tempdir()` to rolling my own temporary function.
Turns out R's tempdir function depends on the system to supply the variables and the parameters.
Depending on which system you are on - you might find these variables inconvienient.

One major inconvinience I ran into was that the size of the tempdir is small enough to cause me problems.

As such I am rolling my own.