Changes:-
ergonomics_functions.r:- Added an ergonomics function `window_unit_func` that makes it easy to check user supplied values for window and convert it from kilobases to bps if need be.

panvar_gwas_tagsnp_snpeff.r :- Made a change that lets the user supply the window value in KB

temporary:- Syntactic sugar for the functions that makes the default value in a better programmatic way. And starting the process of converting the code to let the user supply custom output paths.